Simplify code~

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Hardware.java
@@ -45,7 +45,7 @@ final class Hardware
     static final double GRABBER_GRABBED = 1;
     static final int TETRIX_TICKS_PER_REVOLUTION = 1440;
     static final double TETRIX_TICKS_PER_TURN_DEGREE = 2681d / 300d;
-    static final double LIFTER_TOP_LIMIT = 4 * TETRIX_TICKS_PER_REVOLUTION;
+    static final int LIFTER_TOP_LIMIT = 4 * TETRIX_TICKS_PER_REVOLUTION;
     static final double RELIC_ARM_TOP_LIMIT = 0.5 * TETRIX_TICKS_PER_REVOLUTION;
     static final double JEWEL_ARM_EXTENDED = 0;
     static final double JEWEL_ARM_HALF_EXTENDED = 0.25;
